fix(leave): days 欄位改 NUMERIC(5,2) 支援半天請假
問題:leave_requests.days 原為 INTEGER、但 lib/leave/request-flow.js
寫入 hours/8 為小數、prod 收到 invalid input syntax for type integer "0.5"
任何 hours 非 8 整數倍的請假申請(半天 / 自訂時段)皆無法送出。

修補:widening INTEGER → NUMERIC(5,2)、舊資料無損 cast。
已於 2026-05-05 直接在 prod Supabase SQL Editor 執行、本 commit 補進版控。

- migrations/2026_05_05_leave_days_to_numeric.sql 新增
- supabase_setup.sql days 欄位定義同步更新

+++ supabase_setup.sql
@@ -43,7 +43,8 @@ CREATE TABLE IF NOT EXISTS leave_requests (
   leave_type    TEXT NOT NULL,
   start_date    DATE NOT NULL,
   end_date      DATE NOT NULL,
-  days          INTEGER NOT NULL,
+  -- 2026-05-05: 改 NUMERIC 支援半天 / 自訂時段，見 migrations/2026_05_05_leave_days_to_numeric.sql
+  days          NUMERIC(5,2) NOT NULL,
   reason        TEXT,
   status        TEXT DEFAULT 'pending' CHECK (status IN ('pending','approved','rejected')),
   applied_at    TIMESTAMPTZ DEFAULT NOW(),
